Add PostgreSQL database connection and configuration
This commit adds the necessary code to connect to a PostgreSQL database and configure the connection. It includes the addition of a new environment variable in the .env file for the database URL, as well as the creation of a new function in the Database package to establish the connection. The main.go file has been updated to connect to the database on startup. This change enables the application to communicate with the PostgreSQL database.